LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-preflight-redirect.any.worker-expected.txt

CONSOLE MESSAGE: Preflight response is not successful. Status code: 301
CONSOLE MESSAGE: Preflight response is not successful. Status code: 302
CONSOLE MESSAGE: Preflight response is not successful. Status code: 303
CONSOLE MESSAGE: Preflight response is not successful. Status code: 307
CONSOLE MESSAGE: Preflight response is not successful. Status code: 308

PASS Redirection 301 on preflight failed
FAIL Redirection 301 after preflight failed assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Redirection 302 on preflight failed
FAIL Redirection 302 after preflight failed assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Redirection 303 on preflight failed
FAIL Redirection 303 after preflight failed assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Redirection 307 on preflight failed
FAIL Redirection 307 after preflight failed assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Redirection 308 on preflight failed
FAIL Redirection 308 after preflight failed assert_unreached: Should have rejected: undefined Reached unreachable code
